backend_v2/requirement.txt
fastapi
uvicorn
python-multipart
google-generativeai
vertexai
google-auth
google-genai
google-cloud-aiplatform